Don't require tsconfig to construct a `GlintLanguageServer`
